Improvements to Scalar handling, and implementation of the Blas routines for AutoMatrix. Trying to get a handle on what it's going to take to dynamically dispatch all of these computational routines. This could get unwieldy fast, but the dispatch.hpp macros should help. This does seem to point to a way to go forward on DistMatrix.

+++ include/elemental/core/matrix_impl.hpp
@@ -48,7 +48,7 @@ Int AutoMatrix<Int>::LDim() const
 { return ldim_;  }
 
 template<typename Int> inline
-Int AutoMatrix<Int>::DataSize() const
+size_t AutoMatrix<Int>::DataSize() const
 { return dsize_; }
 
 template <typename Int> inline
